feat: move i18n to its own plugin file
